v6.0.0
See ./mongoose-relay-paginate/docs/Changelog.md for changes relevant parts of that file are also pasted below for convenience.

## v5.0.0 to v6.0.0

### API Breaking Changes

Due to performance reasons anything in this plugin that was dependant on getting Query.countDocuments() or Aggregate.count() (that includes the `$count` aggregation operator) for the whole pagination result had to be omitted. That includes:

- The `count` property in the `RelayResult`'s `pageInfo` which had to use a `Query.countDocuments()`.
- The `toAggregate` method returned from `aggregateRelayPaginate` which had to use an Aggregate `$count` operator

Sorry if you were dependant on this, but our team at work recently saw very bad performance issues due to the previous.

### Alternative solutions to the above

#### pageInfo.count

There is no alternative to the count property that this library used to provide. Also `count` was never spec compliant witht the Relay Connection spec, so it was likely a poor choice to begin with. You could always provide your own external solution to this, but do know that performance will likely be lackluster.

#### toAggregate

Though not the exact same there is now instead a `toNodesAggregate` method. Instead of returning the `RelayResult` it only returns `RelayResult[nodes]` i.e. just the nodes. The following two examples' `result`s are equivalent between the previous version's `toAggregate` and the new `toNodesAggregate`:

**toAggregate**
```ts
const result = await UserModel.aggregateRelayPaginate(
      [{ $sort: { name: 1 } }],
      {}
    )
      .toAggregate()
      .unwind("$nodes")
```

**toNodesAggregate**
```ts
const result = await UserModel.aggregateRelayPaginate(
      [{ $sort: { name: 1 } }],
      {}
    )
      .toNodesAggregate()
```